added "friends list" system.
This is important for safety (i.e., to prevent accidentally sending funds to inaccessible addresses). However it also allows the _senders_ of incoming transfers to be identified easily. This is likely to be an important feature in practice.

It appears to still happen occasionally, unfortunately, that somehow `this.balance` gets corrupted and leads to failures. Not sure though? Maybe look into this.