hide generation of throwaway until _after_ transfer
this improves latency from the perspective of the user. note that this introduces a safety risk: if transfers are "crammed" (that is, if the next one is initiated before the throwaway is re-generated) it's possible that the same address will appear twice. this will easily be solved by some sort of locking / blocking mechanism, and in any case in the future it will make most sense to have an asynchronous worker constantly generating throwaways, which can then be used freely when desired.