FFT even-indexed outputs optimization
because I only use circular shifts of _even_ order, i only require the even-indexed values of the result of circular convolution. thus i can use this optimization https://dsp.stackexchange.com/a/30699 on the inverse transform (instead of just tossing the odd values). this leads to modest but pretty substantial improvements in all measurements of runtime, especially for larger anonsets