fix zero-knowledge vulnerability
This commit mitigates a serious zero-knowledge vulnerability in basic Zether (Sigma-Bullets), which could have allowed an observer to determine b' and b* using only 2^64 work (and far less in some cases). The full amended protocol is implemented in this commit; it's also specified in AnonZether.pdf (the attack and its fix are described on page 6).