more adjustments: separate randomnesses
for each i, j (instead of just for each i), also separate randomness pi for (C_Ln, C_Rn). It is a bit difficult to show that this is _strictly necessary_ for zero-knowledge, but it makes the proofs go through more naturally, and doesn't affect performance much (just have to generate a few extra random scalars). Thus probably the right thing to do. open question to make this exact.

also: the docs now feature (sketches of) security proofs!