now featuring logarithmically-sized proofs.
proof _size_ now grows logarithmically in the size of the anonset. i also have security proof sketches in this new setting, which are significantly more involved. please see docs/AnonZether.pdf for a precise specification of the new protocol, as well as for the proofs. these are only sketches at this point, and are liable to change.