refactor: IP verifier own contract
and deployments are now cheap enough for mainnet use.

the most significant change is that the "g, h, gs, and hs" parameters are now hard-coded into the contract code, which cuts down on storage costs. in particular, each individual deployment can be carried out for 4-6 million gas. this recourse was suggested in the original zether paper, and also by @ibudisteanu (thanks to them).

additionally, the inner product verifier is now a separate contract, used by both the ZetherVerifier and BurnVerifier contracts. this cuts down on repeated code and is more elegant.