refactor: ZSC uses Utils
and all points are of type G1Point, isntead of being "emulated" as bytes32[2]'s.

the advantage of this is essentially that the ZSC can use the "heavy" tools for scalar and point manipulation, and doesn't have to rely on inline assembly. there are a few things that Solidity doesn't properly support yet, so we've had to work around those. but it's an improvement.